Split connection into sync and async.
