+++ ethane/src/connection/non_blocking.rs
@@ -1,0 +1,1 @@
+pub struct Connection;
